backend/requirements.txt
flask
flask-cors
twilio
python-dotenv
anthropic
scikit-learn
numpy
